refactor client :
- create different struct for Go & python
- put Go & Python code in separate file
