+++ codegen/server_resources_gen.go
@@ -68,9 +68,6 @@ func generateServerResources(apiDef *raml.APIDefinition, directory, packageName,
 
 	rs := apiDef.Resources
 
-	if err := checkCreateDir(directory); err != nil {
-		return rds, err
-	}
 	// sort the keys, so we have resource sorted by keys.
 	// the generated code actually don't need it to be sorted.
 	// but test fixture need it
